Remove text-shadow from Slowvote poll options
Summary:
In dark mode, the current `text-shadow` of any potentially displayed unexpected response output offers a nice fuzzy feeling and is a great alternative to the consumption of beverages.

We could either remove it completely or replace `white` with a new `{whitetext}` variable similar to the already existing `{blacktext}` variable. I prefer to keep it simple, so this patch removes the text-shadow.

Refs T15056

Test Plan:
* Go to http://phorge.localhost/settings/panel/display/ and set `Accessibility` to `Dark Mode (Experimental)`
* Create a poll at http://phorge.localhost/vote/create/
* Go to the poll at http://phorge.localhost/V1 and look at the text of the available options

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26153