Replace custom yellow color for table row highlighting by variable
Summary:
Replace hardcolor yellow color `#fdf9e4` for a static highlighted table row with the variable `{$sh-yellowbackground}`. This does not directly improve dark mode but it will later allow to adjust all uses of `{$sh-yellowbackground}` to be suitable for dark mode.

The current yellow highlighting color is static as it is passed by the id URI parameter. It is unrelated to hover behavior.

Note that this is a color change. Previously, the exact yellow color depended on whether the table line is an alternative row or not (every other row is "alternative" for better reading). This patch ignores that for the sake of simplicity. (There is no such differentiation either when hovering over an alt line or non-alt line, the hover color remains `#3c4b5e`.)

Refs T15056

Test Plan:
* Go to http://phorge.localhost/settings/panel/display/ and set `Accessibility` to `Dark Mode (Experimental)`
* Go to http://phorge.localhost/settings/panel/multifactor/ and set up two MFA auth factors for an account
* After adding an auth factor, note the `id` parameter in the URI, highlighting the item in the table; explicitly set these two id parameters when testing

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26156