Use a Unicode character dot on Differential Paths Panel divider
Summary:
Replace the blurry dot PNG rendered on the vertical divider between the Paths Panel and the actual revision in Differential with a Unicode dot character which can easily be rendered in a different color (useful for improved dark mode support).

Also remove the `box-shadow` effect on the right border of the vertical divider. I like clean lines and the shadow made the dot look less horizontally centered.

Refs T15056

Test Plan:
* Create a Differential revision
* Go to http://phorge.localhost/D1
* In the left upper corner, click that arrow thingie to Show Paths Panel
* Look at the vertical divider and spot some dot somewhere on that divider

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26161