Global Search Scope: Replace "Current Application" by application name
Summary:
Be less mysterious about the application we are currently in by replacing "Current Application" with for example "In Differential Only" when we're on a page which belongs to Differential.

Not every user may be aware or understand Phorge's "application" concept, and/or the installation may offer a very limited set of applications only.

There is no good obvious reason why to continue obfuscating what's the current application.

Closes T16167

Test Plan: Go to URIs belonging to applications which have searchable document types, such as http://phorge.localhost/calendar/ or http://phorge.localhost/D1, select the Search Scope dropdown in the upper right corner, see "In $application_name Only" instead of "Current Application"

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16167

Differential Revision: https://we.phorge.it/D26186