Support specifying form fields as readonly
Summary:
Similar to AphrontFormTextAreaControl.

I went with manually specifying the property in each class, as there
are only 5 (AphrontFormTextControl is handled in D26195) and other
types of inputs do not support the readonly property.

Ref T16177

Test Plan: Compare against AphrontFormTextAreaControl. If feeling adventurous, maybe change one of the callers of the input fields to readonly.

Reviewers: O1 Blessed Committers, valerio.bozzolan, aklapper

Reviewed By: O1 Blessed Committers, valerio.bozzolan, aklapper

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16177

Differential Revision: https://we.phorge.it/D26200