a11y: Set aria-labels for top bar dropdown links next to logo/wordmark
Summary:
Links should have a discernible name per https://dequeuniversity.com/rules/axe/4.10/link-name.

This is not the case for the dropdown links (notifications, chat messages, installation setup, account setup) in the top bar next to the Phorge logo/wordmark when a dropdown link has a value of "0" / is empty (means: hidden for installation and account setup, or displayed without a number for notifications and chat messages). Thus add an `aria-label` attribute for screenreader use.

Refs T16072

Test Plan:
* Go to http://phorge.localhost while logged in.
* Inspect the dropdown links in the top bar next to the Phorge logo/wordmark.
* Optionally, run an Accessibility check in Chromium Lighthouse, or try to use Phorge with screenreader software.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16072

Differential Revision: https://we.phorge.it/D26235