Remove orphan Phrequent CSS and icons
Summary:
`.phrequent-tracking-property`, `.phrequent-active` and `.phrequent-on-stack` were added in rP11966756.
Their usage was removed in rP943080a4, which means that also the PNG icons are unused.

Test Plan: Grep the code via `grep --exclude="*.css" -R 'phrequent-' .` and for the PNG filenames, look at the git history, look at Phrequent in the UI.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26245