Replace usages of `wild` with `mixed` or more specific type.
Summary:
The doc type `wild` isn't a recognized phpstan or phpdoc type. The proper type
name for `wild` is `mixed`. For almost every case, it was nontrivial to
identify a more precise type, and so I simply replaced `wild` with `mixed`. In
some cases it was relatively easy and in those cases, either a concrete type or
an inline type alias was used instead of `mixed`.

Test Plan: Run phpstan before and after. See fewer or more specific errors, after.

Reviewers: aklapper, O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26344