Basic emoticons starting with ':' do not suggest nonsense emojis
Summary:
The emoji picker is not opened anymore when you are typing some basic emoticons starting with ':'.

It means that ':D' remains as-is instead of proposing a dagger emoji.

It means that ':P' remains as-is instead of proposing a package emoji. etc.

Closes T16218

Test Plan:
Manually type these emoticons in a remarkup area (e.g. a comment), one per line,
and enjoy the fact that the emoji picker does not propose silly things for them:

    :'(
    :')
    :<
    :>
    :[
    :]
    :|
    :D
    :P

In particular, enjoy `:D` and `:P` which do not propose daggers and packages anymore.

Check for regressions typing these emoticons too, and they still work as-is:

    :-(
    :-)
    :(
    :)

Check for regressions typing some emojis. They still work.
Start typing the "dagger" and "dancer" emoji, you find them:

    :da

Start typing the "package" emoji, you still find it:

    :pa

Reviewers: O1 Blessed Committers, avivey

Reviewed By: O1 Blessed Committers, avivey

Subscribers: tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T16218

Differential Revision: https://we.phorge.it/D26269