PhabricatorTestDataGenerator documentation: minor refactor, minor PHPDoc
Summary:
Minor refactor in loadPhabricatorUserPHID() to do not reinvent the wheel.

Minor improvement of the PHPDoc.

Set methods as final, since nothing overrides them and nothing should.

Adopt the '::class' thing, which is just semantically more appropriate to get exactly the same string.

Test Plan:
Trigger the data generator:

    ./bin/lipsum generate all

It still works.

Optionally, add a `phlog('yup I am executed')` in one of the touched methods,
and repeat the lipsum generation, so you are 100% sure that the code works.

Optionally, run this:

    var_dump(PhabricatorUser::class === 'PhabricatorUser');
    true

So, no regressions. Sigh of relief.

-----

Look for code extending the touched methods:

    grep -FR 'function loadPhabricatorUser' src/

Absolutely nothing else extends them. So, has sense to mark as final.

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26279