lipsum task generator: fix missing 'created this task'
Summary:
The lipsum generator for tasks was bugged, always causing a 'renamed this task' as first feed.

This happened since a manual setTitle() was called, and then the related "title transaction"
was also created, effectively causing a rename.

Just relying on the "title transaction" in the ex line 24 is sufficient.

Closes T16230

Test Plan:
Test the lipsum A/B for tasks:

    ./bin/lipsum generate task

The first feed is not anymore 'renamed' but 'created'.

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T16228

Differential Revision: https://we.phorge.it/D26280