Added a Mentions tab to the Differential Revision View
Summary:
This revision adds a Mentions tab to the revision between the History and Commits tabs

The tab functions exactly like the tab you see on a Maniphest task

When you have revisions that have lots of activity, or, lots of mentions in other places (including tasks, and other revisions) it is hard to see that list in a concise way.   This revision solves this by adding the same tab to a Revision

**This revision does the following: **
- Removes `final` from PHUIPropertyListView so we can extend it
- Created a new class called `PHUIMentionsListView` that extends `PHUIPropertyListView`
- Replaces the code in `ManiphestTaskDetailController` with this new view
- Adds a `Mentions` tab to `DifferentialRevisionViewController`

**Future Work**
We could re-use this in other places in Phorge that support mentions.   Either as a Tab view, or, in some other place in the view.

{F3014887 width=800}

Closes T16003

Test Plan: Made a revision, mentioned it, and saw mentions

Reviewers: O1 Blessed Committers, valerio.bozzolan, avivey

Reviewed By: O1 Blessed Committers, valerio.bozzolan, avivey

Subscribers: avivey, aklapper, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16003

Differential Revision: https://we.phorge.it/D25897