Expose users' MFA status in Project Members list to admins
Summary:
As an admin, I want to make sure that every user has MFA set up who is a member of a project with a restrictive Join Policy.
As far as I know, this is impossible via Dashboard Panels or Advanced Search queries.
Thus expose this MFA information to admins and sort the list of members to first list enabled user accounts with MFA so I can quickly check this as an admin.

Closes T16194

Test Plan:
* Set up two users with Multi-Factor Auth, disable one.
* Go to the Members List of a project as an admin and as a non-admin, look at each user entry and the sorting.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16194

Differential Revision: https://we.phorge.it/D26315