lipsum revisions: fix missing 'created this revision' (adopt transactions)
Summary:
When generating test revisions, the 'created this revision' is not missing anymore,
and the nonsense 'updated this revision to Diff 123' is not shown anymore,
since this is a creation, and not an update.

Basically we avoid to manually call setters, and we use transactions instead.

Closes T16232

Test Plan:
    ./bin/lipsum generate revisions

Enjoy the new 'created this revision'.

You also don't see anymore the nonsense 'updated this revision'.

Enjoy the fake revisions that are now more similar to the real ones.

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, Matthew, Cigaryno

Maniphest Tasks: T16232

Differential Revision: https://we.phorge.it/D26286