User Profile Projects box: Always show View All and number of projects
Summary:
Allow users to easily reach the list of all projects they are a member of, regardless of project status. Currently there is no way to realize that a user is a member of more projects than the active ones listed if the number of active projects is 5 or less, as no "View All" button was shown.

Also always show the number of active projects in brackets for UI consistency.

Also make the "View All" link directly jump to the results list by appending `#R`.

Closes T16259

Test Plan: Go to user profile, look at Projects box title, click "View All" button. Join or leave projects; archive projects the user is a member of.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16259

Differential Revision: https://we.phorge.it/D26332