Fix PHP 8.1 "strlen(null)" exception when using transaction.search
Summary:
`strlen()` was used in Phabricator to check if a generic value is a non-empty string.
This behavior is deprecated since PHP 8.1. Phorge adopts `phutil_nonempty_string()` as a replacement.

Note: this may highlight other absurd input values that might be worth correcting
instead of just ignoring. If phutil_nonempty_string() throws an exception in your
instance, report it to Phorge to evaluate and fix that specific corner case.

This is not visible in the web user interface, only in the PHP error log.

```
strlen(): Passing null to parameter #1 ($string) of type string is deprecated at [TransactionSearchConduitAPIMethod.php:273]
```

Test Plan: * Call Conduit method `transaction.search` with a differential objectIdentifier (e.g. D1)

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26362