Dark Mode: Make lightblueborder slightly lighter for better contrast
Summary:
Make `{$lightblueborder}` a wee bit lighter. This color is used in may places, e.g. for border lines of boxes, so we should not make a larger change, but I believe it is a contrast improvement for circled icons. See screenshots attached in the revision.

Refs T15056

Test Plan:
* Go to http://phorge.localhost/settings/panel/display/ and set `Accessibility` to `Dark Mode (Experimental)`
* Go to http://phorge.localhost/uiexample/view/PHUIIconExample/ and in the "Circles" section, look at both the box border color and the first six icons.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26371