Improve Dark Mode for exiled files behind curtain
Summary:
Use more suited colors (and the variables actually meant for these two elements) for the "File Not Attached" red string in a red box.
Still not perfect, but good enough. (Unfortunately red is red, both in Dark Mode and Default Mode.)

Refs T15056

Test Plan: Go to a task which has a referenced file which is not attached, look at "Referenced Files" in the sidebar.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26381