UIExamples: Fix missing styling
Summary:
This is a regression from ce3c11dc8685.
Previously, the whole page content was part of the navigation,
which itself was set to be a child of the page container.
This caused several styling issues and was fixed. However, the
page content was still a child of the navigation and navigation
doesn't render included styling.

Closes T16283

Test Plan:
* Look at http://phorge.localhost/uiexample/view/PHUIBadgeExample/
* Look at http://phorge.localhost/uiexample/view/PHUILeftRightExample/

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16283

Differential Revision: https://we.phorge.it/D26389