List involved users first in autocomplete proposals in Maniphest
Summary:
Trying to mention a specific user in a task when writing a comment and using the `@username` typeahead functionality in a Phorge installation with tens of thousands of users, currently the autocomplete proposals are an alphabetical list which displays the first five users out of all and any users in that instance.
Thus the more users there are in the instance, the more characters you have to type to find and mention a specific user, on average.
However, often you want to mention a user who has already been involved in the task as an author, assignee, os subscriber.

Thus list those users first who have already been involved with the task when populating the autocomplete proposals for usernames.

This inevitably requires context by letting the PhabricatorRemarkupControl know about the surrounding object (a task) in which it is displayed.

This patch only implements the functionality for Maniphest tasks. It likely makes sense to implement `getInvolvedUsers()` also for other storage objects.

Closes T16126

Test Plan:
* Create several users with the same first letter "a" in their username
* Set some of these users as author, owner, subscribers of a task
* Go to the task comment field and type "@a"
* Look at the sorting in the list of autocomplete proposals in the populated dropdown
* Try with a different first letter not involved in the task, see no difference in behavior
* Try the same with a task which has only a task author but no assignee/owner and no subscribers, still works as expected
* Try the same with an object which does not implement this (e.g. a Differential revision), type ahead and see no difference in behavior and no errors in the browser console

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16126

Differential Revision: https://we.phorge.it/D26395