Fix Dark Mode in Table of Contents of Differential revisions
Summary:
There can be selected, focused, and hovered items in the tree view of a revision.
Make the last two use actual background colors instead of random other stuff.
Also make the hover color for files in the tree consistent with the four section headers in the sidebar.

Note that this is completely unrelated to the Table of Contents in Phriction documents.

Refs T15056

Test Plan:
* Go to http://phorge.localhost/settings/panel/display/ and set `Accessibility` to `Dark Mode (Experimental)`
* Go to a Differential revision like http://phorge.localhost/D1
* Open the sidebar on the left
* Hover over files and the four headings
* Click on a file in the tree
* Scroll in the main view
* Hover over files
* See three different background colors for items, with colors which are acceptable in both Default mode and Dark mode

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15056

Differential Revision: https://we.phorge.it/D26402