Display each setting's config source on All Settings page
Summary:
As an admin, I want to understand which setting is loaded from which configuration source in order to better sync our staging and production instances of Phorge and e.g. transfer useful settings from the Database or Local Source to the Config File source shared between instances.
This is much easier when the config source for a setting is shown in the list of All Settings instead of being buried on each subpage.

Closes T16319

Test Plan:
Go to http://phorge.localhost/config/settings/all/ and look at the column on the right
Open a setting and verify that the source which was displayed in the list is the same as under "Current Configuration"

Reviewers: O1 Blessed Committers, mainframe98, avivey

Reviewed By: O1 Blessed Committers, mainframe98, avivey

Subscribers: avivey, mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16319

Differential Revision: https://we.phorge.it/D26452