Personal API Token field: Auto-select on click
Summary:
When looking at an API Token you often probably want to copy it. While buttons are one way (that's D26222), clicking into the text field is another user pattern.
Highlight the entire string in the text field when doing so, as the "Clone" URIs in Diffusion have the same behavior.

Closes T16335

Test Plan:
* Go to http://phorge.localhost/settings/panel/apitokens/ and Generate Token
* Click on the listed Token
* In the "View API Token" dialog, click in the "Token" field

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16335

Differential Revision: https://we.phorge.it/D26491