Pluralize all the things!
Summary:
(Well, not all of them, but many of them - add pluralization to most strings that need it). See T16378

Also remove the `(s)` plural hack from a few places where it's logically impossible for there to only be one of the given thing, and fix a few random typos, while I'm at it.

Test Plan:
Run P60 and see fewer errors.

Read the code. Try to get some of the strings to show up in the interface but some of them seem to be unfindable.

Reviewers: O1 Blessed Committers, valerio.bozzolan, avivey

Reviewed By: O1 Blessed Committers, valerio.bozzolan, avivey

Subscribers: avivey, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26546