Fix wrong proto-English string for actions with no effect
Summary: The %s is the number of actions with no effect, not the total number of actions.

Test Plan:
- Switch locale to "raw strings"
- Apply two actions one of which has no effect.
- Without this patch see "Some of your 1 action(s) have no effect" which is wrong because you have 2 total actions.
- With this patch see "1 of your actions has no effect".
- Switch locale to US english and see the same things with and without the patch

Note that it's logically impossible for you to be making only one action here; at least one must have no effect (for the message to make sense), and at least one must have some effect (or else it prints a different message).

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26575