CSS: Remove inconsistent phui-timeline-shell.anchor-target box border
Summary:
Show no border at all instead of only on the right side, for the sake of consistency.

Closes T16385

Test Plan:
Go to an anchor link in an object like a Maniphest task, look at the borders of that highlighted box.

Test it with the Dark theme, and normal theme.

Reviewers: O1 Blessed Committers, valerio.bozzolan, mainframe98

Reviewed By: O1 Blessed Committers, valerio.bozzolan, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16385

Differential Revision: https://we.phorge.it/D26593